+++ gimmemotifs/commands/__init__.py
@@ -1,13 +1,14 @@
-import pkgutil
-import os
-
-dirname = os.path.split(__file__)[0]
-
-level = 0
-
-# Dynamically load all commands
-for _importer, cmdname, _ in pkgutil.iter_modules([dirname]):
-    m = __import__(
-        "{0}.{1}".format(__name__, cmdname), globals(), locals(), [cmdname], level
-    )
-    globals()[cmdname] = getattr(m, cmdname)
+from . import (  # noqa: F401
+    motifs,
+    scan,
+    maelstrom,
+    match,
+    logo,
+    cluster,
+    background,
+    threshold,
+    location,
+    diff,
+    prediction,
+    motif2factors,
+)
