+++ gimmemotifs/stats.py
@@ -86,10 +86,12 @@ def calc_stats_iterator(
             if func.input_type == "pos":
                 remove_stats.append(s)
         if len(remove_stats) != 0:
-            logger.warn(
+            logger.warning(
                 "Cannot calculate stats that require position from table of motif scores."
             )
-            logger.warn(f"Skipping the following statistics: {', '.join(remove_stats)}")
+            logger.warning(
+                f"Skipping the following statistics: {', '.join(remove_stats)}"
+            )
             stats = [s for s in stats if s not in remove_stats]
 
     if isinstance(motifs, Motif):
@@ -370,7 +372,7 @@ def write_stats(stats, fname, header=None):
                     )
                 )
             else:
-                logger.warn(
+                logger.warning(
                     "No stats for motif {0}, skipping this motif!".format(motif)
                 )
             # motifs.remove(motif)
